fix some compiler warnings

+++ src/outputs/output_flexiarp.cpp
@@ -15,7 +15,7 @@
         LinkedList<MenuItem*> *items = menu->get_page(menu->get_number_pages()-1)->items;
         SubMenuItemBar *midi_settings_bar = nullptr;
 
-        for (int i = 0 ; i < items->size() ; i++) {
+        for (size_t i = 0 ; i < items->size() ; i++) {
             MenuItem *item = items->get(i);
             if (strcmp(item->label, "MIDI settings")==0) {
                 midi_settings_bar = static_cast<SubMenuItemBar*>(item);
